Вынес LikedCoordinator в отдельный файл
