+++ Navigation/Coordinators/BaseCoordinator.swift
@@ -34,6 +34,10 @@ protocol FeedBaseCoordinator: Coordinator {
     func showDetailScreen()
 }
 
+protocol LikedBaseCoordinator: Coordinator {
+    func showLikedScreen()
+}
+
 protocol MainBaseCoordinator: Coordinator {
     var profileModel: ProfileViewModel { get set }
     var profileCoordinator: ProfileBaseCoordinator { get }
