+++ Navigation/Coordinators/MainCoordinator.swift
@@ -74,28 +74,3 @@ class MainCoordinator: MainBaseCoordinator {
     }
 }
 
-protocol LikedBaseCoordinator: Coordinator {
-    func showLikedScreen()
-}
-
-class LikedCoordinator: LikedBaseCoordinator {
-    
-    var profileModel: ProfileViewModel
-    var parentCoordinator: MainBaseCoordinator?
-    
-    lazy var rootViewController: UIViewController = UIViewController()
-    
-    init (profileModel: ProfileViewModel){
-        self.profileModel = profileModel
-    }
-    
-    func start() -> UIViewController {
-        let likedViewController = LikedPostsViewController(profileViewModel: profileModel)
-        likedViewController.view.backgroundColor = .systemBackground
-        return UINavigationController(rootViewController: likedViewController)
-    }
-    
-    func showLikedScreen() {
-        
-    }
-}
